fix program counter 'ce' and 'ie' issues
remove a faulty short-circuit in the clock-gating circuitry that occured
when input-enable was high. before replacing the old transistor models
in 'tri-state-buffer', this short circuit would not affect the output of
the program counter. hence, the issue didn't appear until now.
my guess is that it was masked by the increased driving capability of
the previous transistors used in the tri-state buffer. even though the
tri-state buffer output was being shorted to ground, its output was
still strong enough to be registered as a logic high by the rest of the
circuit. this fix involves replacing the tri-state logic with a mux-like
circuit that selects between the different inputs depending on the 'ie'
and 'ce' signals. also add some NOTES in the .cmp file about the various
gotchas when interfacing with the program counter.