add feedback to 'dt flip-flop'
when both the 'ie' and 't' control signals are inactive, the output of
the flip-flop is fed back into the input. this prevents the case where
momentarily activating the 't' signal causes the latch to toggle on the
next rising edge, even though 't' was only active for a brief moment.
now, unless 't' or 'ie' are active /at/ the rising edge (and for some
small hold time after), the dt flip-flop won't change its state.